requirements.txt
# Core dependencies
jsonschema>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
regex>=2023.0.0
httpx>=0.25.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Tracing/Observability (PoC-3)
langfuse>=2.0.0

# Optional evaluation engine (PoC-2)
# deepeval>=0.21.0

# Development/Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0